# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Learning Platform API",
        "description": "API for Learning Platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Bio description"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/avatar.jpg"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}/reviews": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user reviews",
                "operationId": "13897806b073f908fa1493c420bb49fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}/certificates": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user certificates",
                "operationId": "a65c5639324c83f4a510a99cab8693f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user certificates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{id}/enrollments": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user enrollments",
                "operationId": "56470dcb03a1fddba46d2680d348f65e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user enrollments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "student_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "course_id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "ACTIVE",
                                                    "COMPLETED",
                                                    "DROPPED"
                                                ]
                                            },
                                            "progress": {
                                                "type": "number",
                                                "format": "decimal"
                                            },
                                            "completed_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "nullable": true
                                            },
                                            "enrolled_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "course": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/filter": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Filter users by role",
                "operationId": "1bc5eb8de7fa0d8e0bb675c85139bf81",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "student",
                            "enum": [
                                "student",
                                "instructor",
                                "admin"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered list of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/avatar.jpg",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "example": "Software developer with 5 years experience",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "STUDENT",
                            "TEACHER",
                            "ADMIN"
                        ],
                        "example": "STUDENT"
                    },
                    "auth_provider": {
                        "type": "string",
                        "enum": [
                            "EMAIL",
                            "GOOGLE",
                            "FACEBOOK",
                            "GITHUB"
                        ],
                        "example": "EMAIL"
                    },
                    "is_email_verified": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Course": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction to Programming"
                    },
                    "description": {
                        "type": "string",
                        "example": "Learn the basics of programming"
                    },
                    "slug": {
                        "type": "string",
                        "example": "introduction-to-programming"
                    },
                    "thumbnail": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/thumbnail.jpg",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PENDING",
                            "PUBLISHED",
                            "ARCHIVED"
                        ],
                        "example": "PUBLISHED"
                    },
                    "level": {
                        "type": "string",
                        "enum": [
                            "BEGINNER",
                            "INTERMEDIATE",
                            "ADVANCED",
                            "EXPERT"
                        ],
                        "example": "BEGINNER"
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal",
                        "example": 99.99,
                        "nullable": true
                    },
                    "duration": {
                        "type": "integer",
                        "example": 30,
                        "nullable": true
                    },
                    "passing_score": {
                        "type": "integer",
                        "example": 70
                    },
                    "teacher_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "teacher": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token (without 'Bearer ' prefix)",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}